// modules/listing/client-react/graphql/__generated__/canUserReview.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: canUserReview
// ====================================================

export interface canUserReview {
  canUserReview: boolean | null;
}

export interface canUserReviewVariables {
  listingId?: number | null;
  userId?: number | null;
}