// modules/listing/client-react/graphql/__generated__/addListing.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

import { AddListingInput } from "./../../../../../packages/server/__generated__/globalTypes";

// ====================================================
// GraphQL mutation operation: addListing
// ====================================================

export interface addListing {
  addListing: number | null;
}

export interface addListingVariables {
  input: AddListingInput;
}