// modules/listing/client-react/graphql/__generated__/duplicateListing.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: duplicateListing
// ====================================================

export interface duplicateListing {
  duplicateListing: number | null;
}

export interface duplicateListingVariables {
  id: number;
}