// modules/listing/client-react/graphql/__generated__/editListing.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

import { EditListingInput } from "./../../../../../packages/server/__generated__/globalTypes";

// ====================================================
// GraphQL mutation operation: editListing
// ====================================================

export interface editListing {
  editListing: boolean | null;
}

export interface editListingVariables {
  input: EditListingInput;
}