// modules/listing/client-react/graphql/__generated__/addOrRemoveListingBookmark.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: addOrRemoveListingBookmark
// ====================================================

export interface addOrRemoveListingBookmark {
  addOrRemoveListingBookmark: string | null;
}

export interface addOrRemoveListingBookmarkVariables {
  listingId?: number | null;
  userId?: number | null;
}