// modules/listing/client-react/graphql/__generated__/shareListingByEmail.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

import { ShareListingByEmailInput } from "./../../../../../packages/server/__generated__/globalTypes";

// ====================================================
// GraphQL mutation operation: shareListingByEmail
// ====================================================

export interface shareListingByEmail {
  shareListingByEmail: boolean | null;
}

export interface shareListingByEmailVariables {
  input?: ShareListingByEmailInput | null;
}